+++ tests/terraform/runner/test_plan_runner.py
@@ -25,13 +25,25 @@ from checkov.terraform.plan_runner import Runner, resource_registry
 @parameterized_class([
    {"db_connector": NetworkxConnector},
    {"db_connector": IgraphConnector},
-    {"db_connector": RustworkxConnector},
+   {"db_connector": RustworkxConnector},
 ])
 class TestRunnerValid(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         cls.orig_checks = deepcopy(resource_registry.checks)
         cls.db_connector = cls.db_connector
+
+    def test_py_graph_check(self):
+        if not self.db_connector == IgraphConnector:
+            return
+        current_dir = os.path.dirname(os.path.realpath(__file__))
+        valid_dir_path = current_dir + "/resources/py_graph_check_tf_plan"
+        valid_dir_path_for_external_check = current_dir + '/py_check_tf_plan'
+        runner = Runner(db_connector=self.db_connector())
+        checks_allowlist = ['CKV_AWS_99999']
+        report = runner.run(root_folder=valid_dir_path, external_checks_dir=[valid_dir_path_for_external_check],
+                            runner_filter=RunnerFilter(framework=["terraform_plan"], checks=checks_allowlist))
+        assert len(report.passed_checks) == 3
 
     def test_runner_two_checks_only(self):
         current_dir = os.path.dirname(os.path.realpath(__file__))